wip(Liquid.Cache, Liquid.Domain: Unification of  the base projects in the Liquid.Core project.
