+++ src/Liquid.Cache.NCache/Liquid.Cache.NCache.csproj
@@ -30,7 +30,7 @@
 
   <ItemGroup>
     <PackageReference Include="Alachisoft.NCache.SDK" Version="5.3.4" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />
     <PackageReference Include="NCache.Microsoft.Extensions.Caching" Version="5.3.3" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.6" />
