+++ src/Liquid.Core/Extensions/DependencyInjection/IServiceCollectionExtension.cs
@@ -1,6 +1,10 @@
 ﻿using Liquid.Core.Implementations;
 using Liquid.Core.Interfaces;
+using Liquid.Core.PipelineBehaviors;
+using MediatR;
 using Microsoft.Extensions.DependencyInjection;
+using System.Reflection;
+using FluentValidation;
 
 namespace Liquid.Core.Extensions.DependencyInjection
 {
@@ -27,6 +31,27 @@ namespace Liquid.Core.Extensions.DependencyInjection
                 services.AddScoped<ILiquidCache, LiquidCache>();
 
             return services;
+        }
+
+        /// <summary>
+        /// Injects mediator handler, validator and Liquid native telemetry for handlers.
+        /// </summary>
+        /// <param name="services">Extended service collection.</param>
+        /// <param name="withTelemetry">Indicates if method should inject Liquid Telemetry Behavior.</param>
+        /// <param name="withValidators">Indicates if method should inject Validators.</param>
+        /// <param name="assemblies">List of assemblies that contains handlers and validators implemantations to be injected.</param>
+        public static void AddLiquidHandlers(this IServiceCollection services, bool withTelemetry, bool withValidators, params Assembly[] assemblies)
+        {
+            if (withValidators)
+            {
+                services.AddTransient(typeof(IPipelineBehavior<,>), typeof(LiquidValidationBehavior<,>));
+                services.AddValidatorsFromAssemblies(assemblies);
+            }
+
+            if (withTelemetry)
+                services.AddTransient(typeof(IPipelineBehavior<,>), typeof(LiquidTelemetryBehavior<,>));
+
+            services.AddMediatR(cfg => cfg.RegisterServicesFromAssemblies(assemblies));
         }
 
     }
