--- src/Liquid.Domain/Liquid.Domain.csproj
@@ -1,35 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <PackageId>Liquid.Domain</PackageId>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Authors>Avanade Brazil</Authors>
-    <Company>Avanade Inc.</Company>
-    <Product>Liquid - Modern Application Framework</Product>
-    <Copyright>Avanade 2019</Copyright>
-    <PackageProjectUrl>https://github.com/Avanade/Liquid.Domain</PackageProjectUrl>
-    <PackageIcon>logo.png</PackageIcon>
-	<Version>6.0.1</Version>
-    <ProjectGuid>{CE75EB47-68D9-46F5-BF91-60D39E52CC54}</ProjectGuid>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <IsPackable>true</IsPackable>
-    <DebugType>Full</DebugType>
-    <Description>The Liquid.Domain contains an abstract class to be inherited by the domain classes of your application.</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentValidation" Version="11.4.0" />
-    <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.4.0" />
-    <PackageReference Include="Liquid.Core" Version="6.0.1" />
-    <PackageReference Include="MediatR" Version="12.1.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\..\logo.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
-
-</Project>
