+++ test/Liquid.Core.Tests/CommandHandlers/Test2/Test2Command.cs
