+++ test/Liquid.Core.Tests/Core/LocalizationTest.cs
@@ -1,4 +1,4 @@
-﻿namespace Liquid.Core.UnitTests
+﻿namespace Liquid.Core.Tests.Core
 {
     public class LocalizationTest
     {
