+++ test/Liquid.Core.Tests/Domain/RequestHandlerTest.cs
@@ -1,7 +1,6 @@
 ﻿using FluentValidation;
-using Liquid.Core.Extensions;
-using Liquid.Domain.Extensions.DependencyInjection;
-using Liquid.Domain.PipelineBehaviors;
+using Liquid.Core.Extensions.DependencyInjection;
+using Liquid.Core.PipelineBehaviors;
 using Liquid.Domain.Tests.CommandHandlers.Test1;
 using Liquid.Domain.Tests.CommandHandlers.Test2;
 using MediatR;
@@ -11,12 +10,12 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Logging.Console;
 using NSubstitute;
-using NUnit.Framework;
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
+using Xunit;
 
-namespace Liquid.Domain.Tests.TestCases
+namespace Liquid.Core.Tests.Domain
 {
     /// <summary>
     /// Base Command Handler Test Class.
@@ -24,7 +23,6 @@ namespace Liquid.Domain.Tests.TestCases
     /// <seealso>
     ///     <cref>Liquid.Test.Base.TestTemplateContext{Liquid.Domain.Tests.TestEntities.TestCommandHandler}</cref>
     /// </seealso>
-    [TestFixture]
     [ExcludeFromCodeCoverage]
     public class RequestHandlerTest
     {
@@ -32,8 +30,8 @@ namespace Liquid.Domain.Tests.TestCases
         private ILogger<LiquidTelemetryBehavior<Test1Command, Test1Response>> _logger = Substitute.For<ILogger<LiquidTelemetryBehavior<Test1Command, Test1Response>>>();
         private ILogger<LiquidTelemetryBehavior<Test2Command, Test2Response>> _logger2 = Substitute.For<ILogger<LiquidTelemetryBehavior<Test2Command, Test2Response>>>();
 
-        [SetUp]
-        protected void EstablishContext()
+
+        public RequestHandlerTest()
         {
             var services = new ServiceCollection();
             services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, ConsoleLoggerProvider>());
@@ -54,13 +52,7 @@ namespace Liquid.Domain.Tests.TestCases
             _serviceProvider = services.AddLogging().BuildServiceProvider();
         }
 
-        [TearDown]
-        protected void TestCleanup()
-        {
-            _serviceProvider = null;
-        }
-
-        [Test]
+        [Fact]
         public async Task Test_WhenCommandHasntValidator_Sucess()
         {
             var mediator = _serviceProvider.GetRequiredService<IMediator>();
@@ -70,10 +62,10 @@ namespace Liquid.Domain.Tests.TestCases
 
             _logger.Received(2);
 
-            Assert.IsNotNull(response);
+            Assert.NotNull(response);
         }
 
-        [Test]
+        [Fact]
         public async Task Test_WhenValidatorPassed_Sucess()
         {
             var mediator = _serviceProvider.GetRequiredService<IMediator>();
@@ -81,14 +73,14 @@ namespace Liquid.Domain.Tests.TestCases
             using var scopedTransaction2 = _serviceProvider.CreateScope();
             var response2 = await mediator.Send(new Test2Command { Id = 1 });
 
-            Assert.IsNotNull(response2);
+            Assert.NotNull(response2);
             _logger2.Received(2);
         }
 
-        [Test]
+        [Fact]
         public void Test_WhenValidatorThrowError_ThowException()
         {
-            Assert.ThrowsAsync<ValidationException>( async () =>
+            Assert.ThrowsAsync<ValidationException>(async () =>
             {
                 var mediator = _serviceProvider.GetRequiredService<IMediator>();
                 using var scopedTransaction = _serviceProvider.CreateScope();
